Begin process of adding space level background decoration, change html background
