+++ index.html
@@ -12,25 +12,6 @@
 
 <body>
 
-    <div id="sun-emoji">
-        <img id="sun" src="assets/scenery/sun.png">
-    </div>
-
-    <div id="cloud-emoji-1">
-        <img class="cloud" src="assets/scenery/cloud.png">
-    </div>
-    <div id="cloud-emoji-2">
-        <img class="cloud" src="assets/scenery/cloud.png">
-    </div>
-    <div id="cloud-emoji-3">
-        <img class="cloud" src="assets/scenery/cloud.png">
-    </div>
-    <div id="cloud-emoji-4">
-        <img class="cloud" src="assets/scenery/cloud.png">
-    </div>
-
-    
-
     <!-- Canvas hosts the actual gameplay as rendered with Kaboom -->
     <canvas id="game_canvas"></canvas>
